Code refractoring for loggin module. Added metrics, and param fetching
