+++ packages/sdk/pureml/components/log/__init__.py
@@ -1,2 +1,2 @@
-from pureml.components import get_org_id, get_token
+from pureml.components import get_org_id, get_token, convert_values_to_string
 from .log import log
