+++ packages/sdk/pureml/components/log/image.py
