+++ packages/sdk/pureml/components/log/log.py
@@ -1,10 +1,17 @@
 import typing
 import numpy as np
 import matplotlib.pyplot as plt
-from .. import metrics as pure_metrics
-from .. import params as pure_params
+from . import metrics as pure_metrics
+from . import params as pure_params
 from . import figure as pure_figure
 from pureml.utils.version_utils import parse_version_label
+from pureml.schema import PathSchema, BackendSchema
+import requests
+from urllib.parse import urljoin
+from . import get_org_id, get_token
+
+path_schema = PathSchema().get_instance()
+backend_schema = BackendSchema().get_instance()
 
 
 def log(label: str = None, metrics=None, params=None, step=1, **kwargs):
